+++ .gitignore
@@ -1,5 +1,4 @@
-node_modules
 *.log
 .DS_Store
-stats.html
-.npmrc
+node_modules
+dist
