--- babel.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-    presets: [
-      [
-        '@babel/preset-env',
-        {
-          targets: {
-            node: 'current',
-          },
-        },
-      ],
-    ],
-  };
