+++ package.json
@@ -1,61 +1,42 @@
 {
-  "name": "boel",
-  "version": "1.0.4",
-  "description": "Form Data validation library",
-  "main": "dist/boel.ssr.js",
-  "module": "dist/boel.esm.js",
-  "unpkg": "dist/boel.min.js",
-  "types": "dist/src/entry.d.ts",
-  "typings": "dist/src/entry.d.ts",
-  "files": [
-    "dist/*"
-  ],
-  "scripts": {
-    "clean": "rimraf dist",
-    "prebuild": "yarn clean",
-    "build": "cross-env NODE_ENV=production rollup --extend --config build/rollup.config.js",
-    "watch": "cross-env NODE_ENV=development rollup --extend --config build/rollup.config.js --watch",
-    "release": "yarn publish --no-git-tag-version",
-    "test": "jest"
-  },
-  "keywords": [
-    "Form Validation",
-    "Javascript Form Validation"
-  ],
-  "author": "Prasanth Janardhanan",
+  "version": "1.0.6",
   "license": "MIT",
-  "dependencies": {
-    "expr-eval": "^2.0.2"
+  "main": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "engines": {
+    "node": ">=10"
   },
+  "scripts": {
+    "start": "tsdx watch",
+    "build": "tsdx build --format umd",
+    "test": "tsdx test",
+    "lint": "tsdx lint",
+    "prepare": "tsdx build"
+  },
+  "peerDependencies": {},
+  "husky": {
+    "hooks": {
+      "pre-commit": "tsdx lint"
+    }
+  },
+  "prettier": {
+    "printWidth": 80,
+    "semi": true,
+    "singleQuote": true,
+    "trailingComma": "es5"
+  },
+  "name": "boel",
+  "author": "Prasanth Janardhanan",
+  "module": "dist/boel.esm.js",
   "devDependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "@types/common-tags": "^1.8.0",
-    "@types/jasmine-expect": "^3.8.1",
-    "@types/jest": "^24.0.18",
-    "babel-jest": "^24.8.0",
-    "cross-env": "^5.2.0",
-    "jasmine-expect": "^4.0.3",
-    "jest": "^24.9.0",
-    "rimraf": "^3.0.0",
-    "rollup": "^1.19.4",
-    "rollup-plugin-babel": "^4.3.3",
-    "rollup-plugin-commonjs": "^10.0.2",
-    "rollup-plugin-graph": "^2.0.0",
-    "rollup-plugin-json": "^4.0.0",
-    "rollup-plugin-node-builtins": "^2.1.2",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-peer-deps-external": "^2.2.0",
-    "rollup-plugin-postcss": "^2.0.3",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-terser": "^5.1.1",
-    "rollup-plugin-typescript2": "^0.24.0",
-    "rollup-plugin-visualizer": "^2.7.2",
-    "sass-loader": "^7.2.0",
-    "ts-jest": "^24.0.2",
-    "ts-jest-transformer": "^1.0.0",
-    "ts-node": "^8.4.1",
+    "expr-eval": "^2.0.2",
+    "husky": "^4.2.5",
+    "tsdx": "^0.13.2",
     "tslib": "^2.0.1",
-    "typescript": "^3.6.2"
+    "typescript": "^3.9.7"
   }
 }
