+++ tsconfig.json
@@ -1,24 +1,22 @@
 {
-    "compilerOptions": {
-      
-      "target": "es5",
-      "strict": true,
-      "removeComments": true,
-      "noImplicitAny": true,
-      "preserveConstEnums": true,
-      "sourceMap": true,
-      "inlineSources": true,
-      "declaration": true,
-      "experimentalDecorators":true,
-      "esModuleInterop":true,
-      "resolveJsonModule":true,
-      "types": ["node","jest"],
-      "module": "es2015",
-      "moduleResolution": "node"
+  "include": ["src", "types"],
+  "compilerOptions": {
+    "target": "es5",
+    "module": "es2015",
+    "importHelpers": true,
+    "declaration": true,
+    "sourceMap": true,
+    "rootDir": "./src",
+    "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noImplicitReturns": true,
+    "noFallthroughCasesInSwitch": true,
+    "moduleResolution": "node",
+    "baseUrl": "./",
+    "paths": {
+      "*": ["src/*", "node_modules/*"]
     },
-    
-    
-    "exclude": [
-      "node_modules"
-    ]
+    "esModuleInterop": true
+  }
 }
